Unexport helper and rename unit test package
